Move MockSharedMemory into service directory and create symlink
